+++ sportWear/src/main/java/com/ecommerce/sportWear/service/ProductoService.java
@@ -8,9 +8,9 @@ import com.ecommerce.sportWear.model.Producto;
 public interface ProductoService {
 	
 	public List<Producto> consultarProductosTodos();
-	public Optional<Producto> consultarProductoPorId(Integer idProducto);
+	public Optional<Producto> consultarProductoPorId(Long idProducto);
 	public Producto crearProducto(Producto producto);
 	public Producto actualizarProducto(Producto producto);
-	public void eliminarProducto(Integer idProducto);
+	public void eliminarProducto(Long idProducto);
 	
 }
